pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "knit-decode"
version = "0.1.0"
description = "Minimal stitch-code decoding pipeline for knit QA artifacts"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "Pillow>=10.0.0",
]

[project.scripts]
knit-decode = "knit_decode.cli:main"

[project.optional-dependencies]
train = [
  "torch>=2.1.0",
]

[tool.setuptools.packages.find]
include = ["knit_decode*"]
